+++ R/reduceResultsBatchmark.R
@@ -29,6 +29,8 @@ reduceResultsBatchmark = function(ids = NULL, store_backends = TRUE, reg = batch
   tabs = split(tabs, by = "job.name")
   bmr = mlr3::BenchmarkResult$new()
 
+  no_warning_given = TRUE
+
   for (tab in tabs) {
     job = batchtools::makeJob(tab$job.id[1L], reg = reg)
     bmr_tasks = bmr$tasks
@@ -60,6 +62,16 @@ reduceResultsBatchmark = function(ids = NULL, store_backends = TRUE, reg = batch
     }
 
     results = batchtools::reduceResultsList(tab$job.id, reg = reg)
+
+    if (no_warning_given & length(results) && mlr_reflections$package_version != results[[1]]$learner_state$mlr3_version) {
+      lg$warn(paste(sep = "\n",
+        "The mlr3 version (%s) from one of the trained learners differs from the currently loaded mlr3 version (%s).",
+        "This can lead to unexpected behavior and we recommend installing the package versions used during experiment exectution."),
+        results[[1]]$learner_state$mlr3_version, mlr_reflections$package_version)
+
+      no_warning_given = FALSE
+    }
+
     rdata = mlr3::ResultData$new(data.table(
       task = list(task),
       learner = list(learner),