+++ NEWS.md
@@ -1,5 +1,8 @@
 # mlr3batchmark (development version)
 
+* docs: A warning is now given when the loaded mlr3 version differs from the
+mlr3 version stored in the trained learners
+
 # mlr3batchmark 0.1.1
 
 * feat: `mlr3batchmark` now depends on package `batchtools` to avoid having to load `batchtools` explicitly.